amp-embedly: amp-embedly-key skeleton
