+++ examples/amp-embedly.amp.html
@@ -15,9 +15,11 @@
   <script async src="https://cdn.ampproject.org/v0.js"></script>
 </head>
 <body>
-  <amp-embedly data-key="13afae3503ee432ca36b630a4b4f656d" data-url="https://twitter.com/AMPhtml/status/970787731533189120" layout="responsive" width="150" height="80"></amp-embedly>
-  <amp-embedly data-key="13afae3503ee432ca36b630a4b4f656d" data-url="https://vimeo.com/18150336" layout="responsive" width="150" height="80"></amp-embedly>
-  <amp-embedly data-key="13afae3503ee432ca36b630a4b4f656d" data-url="https://www.youtube.com/watch?v=TX3sFXHwXjo" layout="responsive" width="150" height="80"></amp-embedly>
-  <amp-embedly data-key="13afae3503ee432ca36b630a4b4f656d" data-url="https://www.flickr.com/photos/digitalweb/14964551" layout="responsive" width="150" height="80"></amp-embedly>
+  <amp-embedly-key value="13afae3503ee432ca36b630a4b4f656d" layout="nodisplay"></amp-embedly-key>
+
+  <amp-embedly data-url="https://twitter.com/AMPhtml/status/970787731533189120" layout="responsive" width="150" height="80"></amp-embedly>
+  <amp-embedly data-url="https://vimeo.com/18150336" layout="responsive" width="150" height="80"></amp-embedly>
+  <amp-embedly data-url="https://www.youtube.com/watch?v=TX3sFXHwXjo" layout="responsive" width="150" height="80"></amp-embedly>
+  <amp-embedly data-url="https://www.flickr.com/photos/digitalweb/14964551" layout="responsive" width="150" height="80"></amp-embedly>
 </body>
 </html>
